+++ packages/vsce-api/__tests__/__unit__/ResourceTypes.test.ts
@@ -22,7 +22,8 @@ describe("ResourceType tests", () => {
       "CICSURIMap",
       "CICSTask",
       "CICSPipeline",
-      "CICSWebService"
+      "CICSWebService",
+      "CICSJVMServer",
     ]);
   });