+++ packages/vsce/__tests__/__unit__/doc/meta.jvmserver.test.ts
@@ -1,0 +1,20 @@
+import { JVMServerMeta } from "../../../src/doc/meta/JVMServer.meta";
+import { IJVMServer } from "../../../src/doc/resources/IJVMServer";
+import { Resource } from "../../../src/resources";
+
+describe("JVMServer Meta", () => {
+  let jvmserverMock: Resource<IJVMServer>;
+
+  beforeEach(() => {
+    jvmserverMock = new Resource({
+      eyu_cicsname: "MYREG",
+      name: "JVM1",
+      status: "ENABLED",
+      enablestatus: "ENABLED",
+    });
+  });
+  it("should return icon name", () => {
+    const iconName = JVMServerMeta.getIconName(jvmserverMock);
+    expect(iconName).toEqual(`jvm-server`);
+  });
+});