+++ packages/vsce/__tests__/__unit__/extension.unit.test.ts
@@ -53,7 +53,7 @@ describe("extension", () => {
     expect(Object.keys(returnedAPI.resources)).toHaveLength(1);
 
     expect(returnedAPI.resources.supportedResources).toBeInstanceOf(Array);
-    expect(returnedAPI.resources.supportedResources).toHaveLength(9);
+    expect(returnedAPI.resources.supportedResources).toHaveLength(10);
     expect(returnedAPI.resources.supportedResources).toContain("CICSProgram");
     expect(returnedAPI.resources.supportedResources).toContain("CICSLocalFile");
     expect(returnedAPI.resources.supportedResources).toContain("CICSTask");